+++ libsponge/byte_stream.hh
@@ -9,43 +9,43 @@ class ByteStream {
     bool _error{};
 
     std::deque<char> _buffer = {};
-    size_t _capacity = 0;
+    uint64_t _capacity = 0;
 
     bool _input_ended_flog = false;
 
-    size_t _write_count = 0;
-    size_t _read_count = 0;
+    uint64_t _write_count = 0;
+    uint64_t _read_count = 0;
 
   public:
-    ByteStream(const size_t capacity);
+    ByteStream(const uint64_t capacity);
 
-    size_t write(const std::string &data);
+    uint64_t write(const std::string &data);
 
-    size_t remaining_capacity() const;
+    uint64_t remaining_capacity() const;
 
     void end_input();
 
     void set_error() { _error = true; }
 
-    std::string peek_output(const size_t len) const;
+    std::string peek_output(const uint64_t len) const;
 
-    void pop_output(const size_t len);
+    void pop_output(const uint64_t len);
 
-    std::string read(const size_t len);
+    std::string read(const uint64_t len);
 
     bool input_ended() const;
 
     bool error() const { return _error; }
 
-    size_t buffer_size() const;
+    uint64_t buffer_size() const;
 
     bool buffer_empty() const;
 
     bool eof() const;
 
-    size_t bytes_written() const;
+    uint64_t bytes_written() const;
 
-    size_t bytes_read() const;
+    uint64_t bytes_read() const;
 };
 
 #endif  // SPONGE_LIBSPONGE_BYTE_STREAM_HH
